[Feat] Completed Item Data
