--- Assets/Resources/Json/Fish.json
@@ -1,24 +1,0 @@
-{
-    "fishes":[
-        {"itemID":10100,"fishName":"다슬기","waterType":0,"fishTear":1,"growTime":2,"feedAmount":0.1,"sellPrice":1000,"buyPrice":900},
-        {"itemID":10200,"fishName":"민물새우","waterType":0,"fishTear":1,"growTime":2,"feedAmount":0.1,"sellPrice":1000,"buyPrice":900},
-        {"itemID":10300,"fishName":"미꾸라지","waterType":0,"fishTear":2,"growTime":3,"feedAmount":0.2,"sellPrice":1500,"buyPrice":1350},
-        {"itemID":10400,"fishName":"송어","waterType":0,"fishTear":2,"growTime":3,"feedAmount":0.2,"sellPrice":1500,"buyPrice":1350},
-        {"itemID":10500,"fishName":"뱀장어","waterType":0,"fishTear":3,"growTime":10,"feedAmount":1,"sellPrice":8000,"buyPrice":7200},
-        {"itemID":20100,"fishName":"김","waterType":1,"fishTear":1,"growTime":2,"feedAmount":0.15,"sellPrice":1250,"buyPrice":1125},
-        {"itemID":20200,"fishName":"미역","waterType":1,"fishTear":1,"growTime":2,"feedAmount":0.15,"sellPrice":1250,"buyPrice":1125},
-        {"itemID":20300,"fishName":"우럭","waterType":1,"fishTear":1,"growTime":2,"feedAmount":0.15,"sellPrice":1250,"buyPrice":1125},
-        {"itemID":20400,"fishName":"광어","waterType":1,"fishTear":2,"growTime":3,"feedAmount":0.2,"sellPrice":1750,"buyPrice":1575},
-        {"itemID":20500,"fishName":"굴","waterType":1,"fishTear":2,"growTime":3,"feedAmount":0.2,"sellPrice":1750,"buyPrice":1575},
-        {"itemID":20600,"fishName":"멸치","waterType":1,"fishTear":2,"growTime":3,"feedAmount":0.2,"sellPrice":1750,"buyPrice":1575},
-        {"itemID":20700,"fishName":"명태","waterType":1,"fishTear":3,"growTime":4,"feedAmount":0.25,"sellPrice":2500,"buyPrice":2250},
-        {"itemID":20800,"fishName":"전복","waterType":1,"fishTear":3,"growTime":4,"feedAmount":0.25,"sellPrice":2500,"buyPrice":2250},
-        {"itemID":20900,"fishName":"쥐치","waterType":1,"fishTear":3,"growTime":4,"feedAmount":0.25,"sellPrice":2500,"buyPrice":2250},
-        {"itemID":21000,"fishName":"고등어","waterType":1,"fishTear":4,"growTime":5,"feedAmount":0.4,"sellPrice":4500,"buyPrice":4050},
-        {"itemID":21100,"fishName":"청어","waterType":1,"fishTear":4,"growTime":5,"feedAmount":0.4,"sellPrice":4500,"buyPrice":4050},
-        {"itemID":21200,"fishName":"오징어","waterType":1,"fishTear":4,"growTime":5,"feedAmount":0.4,"sellPrice":4500,"buyPrice":4050},
-        {"itemID":21300,"fishName":"참다랑어","waterType":1,"fishTear":5,"growTime":8,"feedAmount":0.8,"sellPrice":6000,"buyPrice":5400},
-        {"itemID":21400,"fishName":"철갑상어","waterType":1,"fishTear":5,"growTime":8,"feedAmount":0.8,"sellPrice":6000,"buyPrice":5400},
-        {"itemID":21500,"fishName":"랍스터","waterType":1,"fishTear":5,"growTime":8,"feedAmount":0.8,"sellPrice":6000,"buyPrice":5400}
-    ]
-}
