--- Assets/Resources/Json/Fish.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e45795a3fa4474a47a6b9e6f0f402c74
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
