--- Assets/Resources/Json/Recipe.json
@@ -1,35 +1,0 @@
-{
-    "recipes": [
-        {"itemID":10201,"recipeName":"건민물새우","SellPrice":1500},
-        {"itemID":10202,"recipeName":"새우젓","SellPrice":2625},
-        {"itemID":10301,"recipeName":"냉동 미꾸라지","SellPrice":2250},
-        {"itemID":10401,"recipeName":"냉동 송어","SellPrice":2250},
-        {"itemID":20101,"recipeName":"말린 김","SellPrice":1875},
-        {"itemID":20201,"recipeName":"말린 미역","SellPrice":1875},
-        {"itemID":20301,"recipeName":"냉동 우럭","SellPrice":1875},
-        {"itemID":20302,"recipeName":"어묵(우럭)","SellPrice":4922},
-        {"itemID":20401,"recipeName":"냉동 광어","SellPrice":2625},
-        {"itemID":20402,"recipeName":"어묵(광어)","SellPrice":6891},
-        {"itemID":20501,"recipeName":"굴젓","SellPrice":4594},
-        {"itemID":20502,"recipeName":"굴 통조림","SellPrice":8859},
-        {"itemID":20601,"recipeName":"말린 멸치","SellPrice":2625},
-        {"itemID":20602,"recipeName":"멸치젓","SellPrice":4594},
-        {"itemID":20603,"recipeName":"멸치 통조림","SellPrice":7973},
-        {"itemID":20701,"recipeName":"동태","SellPrice":3750},
-        {"itemID":20702,"recipeName":"황태","SellPrice":3750},
-        {"itemID":20703,"recipeName":"명태젓","SellPrice":6563},
-        {"itemID":20704,"recipeName":"어묵(명태)","SellPrice":9844},
-        {"itemID":20801,"recipeName":"진주","SellPrice":50000},
-        {"itemID":20901,"recipeName":"쥐포","SellPrice":3750},
-        {"itemID":21001,"recipeName":"냉동 고등어","SellPrice":6750},
-        {"itemID":21002,"recipeName":"고등어 통조림","SellPrice":22781},
-        {"itemID":21101,"recipeName":"냉동 청어","SellPrice":6750},
-        {"itemID":21102,"recipeName":"수르스트뢰밍","SellPrice":22781},
-        {"itemID":21201,"recipeName":"오징어포","SellPrice":6750},
-        {"itemID":21202,"recipeName":"오징어 젓갈","SellPrice":11813},
-        {"itemID":21301,"recipeName":"냉동 참다랑어","SellPrice":9000},
-        {"itemID":21302,"recipeName":"참치 젓갈","SellPrice":15750},
-        {"itemID":21303,"recipeName":"어묵(참다랑어)","SellPrice":23625},
-        {"itemID":21401,"recipeName":"캐비어 통조림","SellPrice":30375}
-    ]
-}
