--- Assets/Resources/Json/Recipe.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bf415e4b4f294654a988ef0b08c3c259
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
