--- Assets/Scripts/Item/Fish.cs
@@ -1,20 +1,0 @@
-using System;
-
-[Serializable]
-public class Fish 
-{
-    public int itemID;
-    public string fishName;
-    public bool waterType;
-    public int fishTear;
-    public int growTime;
-    public double feedAmount;
-    public int sellPrice;
-    public int buyPrice;
-}
-
-[Serializable]
-public class FishData
-{
-    public Fish[] fishes;
-}
