--- Assets/Scripts/Item/Fish.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bdf8ae697df507f4f984a1236eedbfab
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
