+++ Assets/Scripts/Item/Item.cs
@@ -1,3 +1,4 @@
+using UnityEngine;
 using System;
 
 [Serializable]
@@ -12,6 +13,14 @@ public class Item
     public int sellPrice;
     public int buyPrice;
     public int power;
+
+    public void ShowItemInfo()
+    {
+        Debug.Log(itemID);
+    }
+
+    public void GeneratePower(int increasePower) { power += increasePower; } // Increase Power Value
+    public void DecreasePower(int decreasePower) { power -= decreasePower; } // Decrease Power Value
 }
 
 [Serializable]
