--- Assets/Scripts/Item/PowerPlant.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4026cee36ec7dfc4dba410c9c3c7fed7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
