--- Assets/Scripts/Item/Recipe.cs
@@ -1,15 +1,0 @@
-using System;
-
-[Serializable]
-public class Recipe
-{
-    public int itemID;
-    public string recipeName;
-    public int sellPrice;
-}
-
-[Serializable]
-public class RecipeData
-{
-    public Recipe[] recipes;
-}
