--- Assets/Scripts/Item/Recipe.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2c8cb7146959e6a4c98aeb5e6c026fdf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
